+++ src/label_projection/methods/mlp/script.py
@@ -1,33 +1,59 @@
+import anndata as ad
+from sklearn.neural_network import MLPClassifier
+import sklearn.pipeline
+import sklearn.preprocessing
+import sklearn.decomposition
+import scipy.sparse
+
 ## VIASH START
 par = {
-    'input': '../../../../resources_test/label_projection/pancreas/toy_preprocessed_data.h5ad',
-    'output': 'output.mlplogcpm.h5ad'
+    'input_train': 'resources_test/label_projection/pancreas/dataset_cpm_train.h5ad',
+    'input_test': 'resources_test/label_projection/pancreas/dataset_cpm_test.h5ad',
+    'output': 'output.h5ad',
+}
+meta = {
+    'functionality_name': 'foo',
 }
 ## VIASH END
-resources_dir = "../../../../common/tools/"
-utils_dir = "../../"
-
-import sys
-sys.path.append(resources_dir)
-sys.path.append(utils_dir)
-sys.path.append(meta['resources_dir'])
-import scanpy as sc
-from utils import classifier
-import sklearn.neural_network
-
 
 print("Load input data")
-adata = sc.read(par['input'])
+input_train = ad.read_h5ad(par['input_train'])
+input_test = ad.read_h5ad(par['input_test'])
 
-if "normalization_method" not in adata.uns:
-    print("Warning: trying to predict for a not normalized data")
+print("Set up classifier pipeline")
+def pca_op(adata_train, adata_test, n_components=100):
+    is_sparse = scipy.sparse.issparse(adata_train.X)
 
+    min_components = min(
+        [adata_train.shape[0], adata_test.shape[0], adata_train.shape[1]]
+    )
+    if is_sparse:
+        min_components -= 1
+    n_components = min([n_components, min_components])
+    if is_sparse:
+        pca_op = sklearn.decomposition.TruncatedSVD
+    else:
+        pca_op = sklearn.decomposition.PCA
+    return pca_op(n_components=n_components)
 
-print("Run classifier")
-hidden_layer_sizes = tuple(par['hidden_layer_sizes'])
-max_iter = par['max_iter']
-adata = classifier(adata, estimator=sklearn.neural_network.MLPClassifier, hidden_layer_sizes=hidden_layer_sizes, max_iter=max_iter)
-adata.uns["method_id"] = meta["functionality_name"]
+classifier = MLPClassifier(
+    max_iter=par["max_iter"], 
+    hidden_layer_sizes=tuple(par["hidden_layer_sizes"])
+)
+pipeline = sklearn.pipeline.Pipeline(
+    [
+        ("pca", pca_op(input_train, input_test, n_components=100)),
+        ("scaler", sklearn.preprocessing.StandardScaler(with_mean=True)),
+        ("regression", classifier),
+    ]
+)
 
-print("Write data")
-adata.write(par['output'], compression="gzip")
+print("Fit to train data")
+pipeline.fit(input_train.layers["lognorm"], input_train.obs["label"].astype(str))
+
+print("Predict on test data")
+input_test.obs["label_pred"] = pipeline.predict(input_test.layers["lognorm"])
+
+print("Write output to file")
+input_test.uns["method_id"] = meta["functionality_name"]
+input_test.write_h5ad(par['output'], compression="gzip")