+++ src/label_projection/metrics/f1/script.py
@@ -1,32 +1,42 @@
+from sklearn.metrics import f1_score
+import sklearn.preprocessing
+import anndata as ad
+
 ## VIASH START
 par = {
-    'input': 'ouput.h5ad',
+    'input_prediction': 'resources_test/label_projection/pancreas/dataset_cpm_knn.h5ad',
+    'input_solution': 'resources_test/label_projection/pancreas/dataset_cpm_solution.h5ad',
     'average': 'weighted',
-    'output': 'output.mv.h5ad'
+    'output': 'output.h5ad'
+}
+meta = {
+    'functionality_name': 'f1'
 }
 ## VIASH END
-import sklearn.metrics
-import sklearn.preprocessing
-import scanpy as sc
-
 
 print("Load data")
-adata = sc.read(par['input'])
+input_prediction = ad.read_h5ad(par['input_prediction'])
+input_solution = ad.read_h5ad(par['input_solution'])
 
-print("Get prediction accuracy")
-encoder = sklearn.preprocessing.LabelEncoder().fit(adata.obs["celltype"])
-test_data = adata[~adata.obs["is_train"]]
+assert (input_prediction.obs_names == input_solution.obs_names).all()
 
-test_data.obs["celltype"] = encoder.transform(test_data.obs["celltype"])
-test_data.obs["celltype_pred"] = encoder.transform(test_data.obs["celltype_pred"])
+print("Encode labels")
+encoder = sklearn.preprocessing.LabelEncoder().fit(input_solution.obs["label"])
+input_solution.obs["label"] = encoder.transform(input_solution.obs["label"])
+input_prediction.obs["label_pred"] = encoder.transform(input_prediction.obs["label_pred"])
 
-metrics = sklearn.metrics.f1_score(
-    test_data.obs["celltype"], test_data.obs["celltype_pred"], average=par["average"]
-)
+print("Compute F1 score")
+metric_type = [ "macro", "micro", "weighted" ]
+metric_id = [ "f1_" + x for x in metric_type]
+metric_value = [ f1_score(
+        input_solution.obs["label"], 
+        input_prediction.obs["label_pred"], 
+        average=x
+    ) for x in metric_type ]
 
 print("Store metric value")
-adata.uns["metric_id"] = meta["functionality_name"]
-adata.uns["metric_value"] = metrics
+input_prediction.uns["metric_ids"] = metric_id
+input_prediction.uns["metric_values"] = metric_value
 
 print("Writing adata to file")
-adata.write(par['output'], compression="gzip")
+input_prediction.write_h5ad(par['output'], compression="gzip")